Split routes/oauth2 into multiple modules
- Move the developer parts (creating new OAuth clients) into
  routes/developer_oauth2
- The general OAuth 2.0 code (setting up the oauth2orize library,
  issuing tokens) goes into util/oauth2.js
- APIs that are externally callable go into routes/my_api.js
- The /authorize endpoint, which needs cookie auth + session
  support + CSRF protection, goes into routes/my_oauth2.js
