+++ util/oauth2.js
@@ -9,38 +9,20 @@
 // See COPYING for details
 "use strict";
 
-const Url = require('url');
-const crypto = require('crypto');
-const express = require('express');
 const passport = require('passport');
 const oauth2orize = require('oauth2orize');
-const multer = require('multer');
-const csurf = require('csurf');
 const util = require('util');
-const fs = require('fs');
 const jwt = require('jsonwebtoken');
 
 const BasicStrategy = require('passport-http').BasicStrategy;
 const ClientPasswordStrategy = require('passport-oauth2-client-password').Strategy;
 
 // create OAuth 2.0 server
-var server = oauth2orize.createServer();
+const server = oauth2orize.createServer();
 
 const model = require('../model/oauth2');
-const user = require('../util/user');
 const db = require('../util/db');
-const code_storage = require('../util/code_storage');
-const graphics = require('../almond/graphics');
-const platform = require('../util/platform');
 const secret = require('../util/secret_key');
-const iv = require('../util/input_validation');
-const { BadRequestError, ForbiddenError } = require('../util/errors');
-
-var router = express.Router();
-
-function makeRandom(size = 32) {
-    return crypto.randomBytes(size).toString('hex');
-}
 
 // These strategies are used to authenticate oauth2 clients, not
 // to authenticate users
@@ -186,150 +168,4 @@ server.exchange(oauth2orize.exchange.refreshToken((client, refreshToken, scope, 
     }).catch(done);
 }));
 
-function verifyScope(client, requestedScopes) {
-    if (!requestedScopes)
-        throw new oauth2orize.AuthorizationError("missing scope query parameter", 'invalid_scope');
-
-    const allowed = client.allowed_scopes.split(' ');
-    for (let scope of requestedScopes) {
-        if (allowed.indexOf(scope) < 0)
-            throw new oauth2orize.AuthorizationError("invalid scope", 'invalid_scope');
-    }
-}
-
-function verifyRedirectUrl(client, redirectURI) {
-    if (!redirectURI)
-        throw new oauth2orize.AuthorizationError("invalid redirect_uri", 'unauthorized_client');
-    for (let url of JSON.parse(client.allowed_redirect_uris)) {
-        if (redirectURI.startsWith(url))
-            return;
-    }
-
-    throw new oauth2orize.AuthorizationError("invalid redirect_uri", 'unauthorized_client');
-}
-
-router.get('/authorize', user.requireLogIn, server.authorization((clientID, redirectURI, scope, done) => {
-    db.withClient((dbClient) => {
-        return model.getClients(dbClient, clientID).then((rows) => {
-            if (rows.length < 1)
-                throw new oauth2orize.AuthorizationError("invalid client", 'unauthorized_client');
-            else
-                return rows[0];
-        });
-    }).then((client) => {
-        try {
-            verifyScope(client, scope);
-            verifyRedirectUrl(client, redirectURI);
-            done(null, client, redirectURI);
-        } catch(err) {
-            done(err);
-        }
-    }, (err) => {
-        done(err);
-    });
-}, (client, user, scope, done) => {
-    done(null, !!client.magic_power, { scope });
-}), (req, res, next) => {
-    const parsedUrl = Url.parse(req.oauth2.redirectURI);
-    let origin = parsedUrl.port !== 443 ? (parsedUrl.hostname + ':' + parsedUrl.port) : parsedUrl.hostname;
-
-    return res.render('oauth2_authorize', {
-        page_title: req._("Almond - Authorize Access"),
-        transaction_id: req.oauth2.transactionID,
-        client: req.oauth2.client,
-        scope: req.oauth2.req.scope,
-        origin
-    });
-}, server.errorHandler({ mode: 'indirect' }));
-
-// no need for csurf here, oauth2orize has a uniquely-generated transaction ID that cannot
-// be spoofed with CSRF
-router.post('/authorize', user.requireLogIn, server.decision((req, done) => {
-    done(null, { scope: req.body.scope.split(' ') });
-}));
-
-router.post('/token',
-    passport.authenticate(['oauth2-client-basic', 'oauth2-client-password'], { session: false }),
-    server.token(), server.errorHandler());
-
-async function uploadIcon(clientId, file) {
-    try {
-        try {
-            const image = graphics.createImageFromPath(file.path);
-            image.resizeFit(512, 512);
-            const [stdout,] = await image.stream('png');
-            await code_storage.storeIcon(stdout, 'oauth:' + clientId);
-        } finally {
-            await util.promisify(fs.unlink)(file.path);
-        }
-    } catch(e)  {
-        console.error('Failed to upload icon to S3: ' + e);
-    }
-}
-
-function validateScopes(req, allowedScopes) {
-    if (!Array.isArray(allowedScopes))
-        allowedScopes = [allowedScopes];
-    for (let scope of allowedScopes) {
-        if (typeof scope !== 'string' || !user.OAuthScopes.has(scope))
-            throw new BadRequestError(req._("Invalid scope"));
-    }
-    return allowedScopes;
-}
-
-function validateRedirectUrls(req, urls) {
-    for (let url of urls) {
-        const parsed = Url.parse(url);
-        if (parsed.protocol === null || parsed.hostname === null ||
-            (parsed.protocol !== 'https:' && parsed.hostname !== '127.0.0.1'))
-            throw new BadRequestError(req._("Invalid redirect URI (must be an absolute https: URL)"));
-    }
-    return urls;
-}
-
-router.post('/clients/create', multer({ dest: platform.getTmpDir() }).single('icon'), csurf({ cookie: false }), user.requireLogIn, user.requireDeveloper(),
-    iv.validatePOST({ scope: ['array', '?string'], name: 'string', redirect_uri: 'string' }), (req, res, next) => {
-    const name = req.body.name;
-    let scopes, redirectUrls;
-    try {
-        if (!name)
-            throw new BadRequestError(req._("Name must be provided"));
-        if (!req.file)
-            throw new BadRequestError(req._("Must upload an icon"));
-        scopes = validateScopes(req, req.body.scope);
-
-        if (scopes.indexOf('profile') < 0)
-            scopes.push('profile');
-        scopes.sort();
-
-        if ((req.user.roles & user.Role.ADMIN) === 0 &&
-            scopes.indexOf('user-sync') >= 0)
-            throw new ForbiddenError(req._("user-sync scope is valid only for administrators"));
-
-        redirectUrls = validateRedirectUrls(req, req.body.redirect_uri.split(/ +/));
-    } catch(e) {
-        res.status(400).render('error', { page_title: req._("Almond - Error"),
-                                          message: e });
-        return;
-    }
-
-    db.withTransaction(async (dbClient) => {
-        const clientId = makeRandom(8);
-        const clientSecret = makeRandom();
-        await model.createClient(dbClient, {
-            id: clientId,
-            secret: clientSecret,
-            name: name,
-            owner: req.user.developer_org,
-            allowed_scopes: scopes.join(' '),
-            allowed_redirect_uris: JSON.stringify(redirectUrls)
-        });
-
-        // upload the icon asynchronously to avoid blocking the request
-        uploadIcon(clientId, req.file);
-    }).then(() => {
-        res.redirect(303, '/developers/oauth');
-    }).catch(next);
-});
-
-module.exports = router;
+module.exports = server;
